plotter, change m_DPR from real to int

+++ src/qtgui/plotter.h
@@ -272,7 +272,7 @@ private:
     QPixmap     m_WaterfallPixmap;
     QColor      m_ColorTbl[256];
     QSize       m_Size;
-    qreal       m_DPR{};
+    int         m_DPR{};
     QString     m_HDivText[HORZ_DIVS_MAX+1];
     bool        m_Running;
     bool        m_DrawOverlay;
